# task4/repo-stat/subscriber/internal/adapter/repository/gen/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package generated_db

import (
	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type Subscription struct {
	ID        uuid.UUID
	RepoName  string
	OwnerName string
	CreatedAt pgtype.Timestamptz
}